chore: Integrate Knative and Skaffold in development (#14)
* fix: Make the server start

I discovered this issue whilst integrating Knative to run the server in development. I'll create a separate PR for that, but I thought it'd be good to get this out of the way soon.

Changes tested with `AUTHORITY_VERSION=1.0 node --loader ts-node/esm src/bin/server.ts`.

* chore: Fix exit handling test suite

As in, make sure all tests pass. ESLint is reporting errors on these files.

* wip

* wip

* wip

* docs

* integrate shared workflows

* empty

* empty

* empty

* empty

* empty

* test custom kn install

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* skip explicit cluster creation

* explicitly set k8s version and cluster name

* try minikube

* Revert "try minikube"

This reverts commit c4924003c30b4c46db18a59dc9e83bd0eb7270d3.

* diagnostics

* diagnostics 2

* drop use of `sudo`

* diagnostics

* diagnostics

* try yaml install

* diagnostics

* diagnostics

* diagnostics

* fix yaml

* wait for kourier stuff

* diagnostics

* diagnostics

* delete diagnostics

* empty

* empty

* disable releases