docs: Move Awala endpoint to root of API (#83)
This endpoint requires anonymous requests, so we can't have it under `/orgs/${orgName}` once we require authentication under `/orgs`. Plus, in hindsight, it never really made sense to have it under `/orgs`.

A consequence of this change is that we need a new outgoing message to acknowledge the key import (`MemberPublicKeyImportAck`).

I also took this opportunity to specify the content type of each message.