refactor: Nest routes
This is groundwork for #84, as we need to use [Fastify encapsulation](https://www.fastify.io/docs/latest/Reference/Encapsulation/) to enforce authz.

Note that I also had to merge `configureFastify()` into `makeServer()` as we'll also need to more control over plugins (their execution order and the parameters we pass to them). It never really make sense to separate the two in the first place (I copied them from another project where we have multiple servers, so it makes sense there).