refactor: Nest routes and make `configureFastify()` more flexible (#89)
This is groundwork for #84, as we need to use [Fastify encapsulation](https://www.fastify.io/docs/latest/Reference/Encapsulation/) to enforce authz.

Note that I also had to convert `configureFastify()` into something that just registers a "plugin", as we'll also need to more control over plugins (their execution order and the parameters we pass to them). I'm calling the new function `makeFastify()` for consistency with Fastify's style (for example, to initialise a Fastify _instance_, you call the default export from `fastify`).